oh hey we can have the same test for buffer!
